% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{cytoverse_deps}
\alias{cytoverse_deps}
\title{List all cytoverse dependencies}
\usage{
cytoverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
cytoverse packages.}
}
\description{
List all cytoverse dependencies
}
