% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cytoverse_packages}
\alias{cytoverse_packages}
\title{List all packages in the cytoverse}
\usage{
cytoverse_packages(pkgs = "all", include_self = TRUE)
}
\arguments{
\item{pkgs}{A character specifying which packages to list. "all" will include all cytoverse packages while "core" will include only those core packages
that will be loaded by library(cytoverse)}

\item{include_self}{Include cytoverse in the list?}
}
\value{
a character vector of package names
}
\description{
List all packages in the cytoverse
}
\examples{
\dontrun{
cytoverse_packages()
}
}
