% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cytoverse_remote_version}
\alias{cytoverse_remote_version}
\title{Check available version for a single cytoverse package}
\usage{
cytoverse_remote_version(pkg, repo = "bioconductor", branch = NULL)
}
\arguments{
\item{pkg}{package name (see \code{\link{cytoverse_packages}} for all options)}

\item{repo}{either "bioconductor" or "github", specifying repository source}

\item{branch}{optionally allows specification of a branch within the repository. For Bioconductor, this
corresponds to the version (e.g. "3.11" or "devel") and will default to the value of \code{\link[BiocManager:install]{BiocManager::version()}}. 
For GitHub, this refers to branch from the GitHub repositories and will default to "master".}
}
\value{
a string representing the package version (e.g "4.1.4")
}
\description{
This method will check the currently available version of a specified
cytoverse package from the specified repository source.
}
\examples{
\dontrun{
# Defaults to checking Bioconductor
cytoverse_remote_version("flowWorkspace")

cytoverse_remote_version("flowWorkspace", repo = "github")
}
}
