% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{cytoverse_update_check}
\alias{cytoverse_update_check}
\title{Check available updates for all cytoverse packages}
\usage{
cytoverse_update_check(repo = "bioconductor", pkgs = NULL, branch = NULL)
}
\arguments{
\item{repo}{either "bioconductor" or "github", specifying repository source}

\item{pkgs}{A character vector containing a specific subset of cytoverse packages to check for updates. 
If NULL, all cytoverse packages will be checked. For the full list of packages, see \code{\link{cytoverse_packages}}.}

\item{branch}{optionally allows specification of a branch within the repository. For Bioconductor, this
corresponds to the version (e.g. "3.11" or "devel") and will default to the value of \code{\link[BiocManager:install]{BiocManager::version()}}. 
For GitHub, this refers to branch from the GitHub repositories and will default to "master".}
}
\value{
a length-2 list containing a dataframe summarizing the installed and available package versions and 
a vector of package names to be updated
}
\description{
Check available updates for all cytoverse packages
}
\examples{
\dontrun{
# Defaults to Bioconductor
check_results <- cytoverse_update_check()
check_results$summary
check_results$lto_update
}
}
