% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cytoverse_versions}
\alias{cytoverse_versions}
\title{List installed package versions for all packages in the cytoverse}
\usage{
cytoverse_versions(pkgs = "all", include_self = TRUE, print = TRUE)
}
\arguments{
\item{pkgs}{A character specifying which packages to list. "all" will include all cytoverse packages while "core" will include only those core packages
that will be loaded by library(cytoverse)}

\item{include_self}{Include cytoverse in the list?}

\item{print}{Whether to display the formatted versions or simply return them}
}
\value{
a named character vector of package versions
}
\description{
List installed package versions for all packages in the cytoverse
}
\examples{
\dontrun{
cytoverse_versions()
}
}
