% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cytoverse_packages}
\alias{cytoverse_packages}
\title{List all packages in the cytoverse}
\usage{
cytoverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include cytoverse in the list?}
}
\description{
List all packages in the cytoverse
}
\examples{
cytoverse_packages()
}
