% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{cytoverse_update}
\alias{cytoverse_update}
\title{Update cytoverse packages}
\usage{
cytoverse_update(
  repo = "bioconductor",
  pkgs = NULL,
  branch = NULL,
  check_versions = TRUE,
  upgrades_approved = FALSE,
  ...
)
}
\arguments{
\item{repo}{either "bioconductor" or "github", specifying repository source}

\item{pkgs}{A character vector containing a specific subset of cytoverse packages to check for updates. 
If NULL, all cytoverse packages will be checked. For the full list of packages, see \code{\link{cytoverse_packages}}.}

\item{branch}{optionally allows specification of a branch within the repository. For Bioconductor, this
corresponds to the version (e.g. "3.11" or "devel") and will default to the value of \code{\link[BiocManager:install]{BiocManager::version()}}. 
For GitHub, this refers to branch from the GitHub repositories and will default to "master".}

\item{check_versions}{a logical indicating whether the package versions should be checked to determine which packages to update. This defaults to TRUE, but if switched
to FALSE, installation will default to checking commit SHA. This can be used to force updates for commits between version increments.}

\item{upgrades_approved}{a logical indicating whether the user should be asked to approve/disapprove the proposed package updates. Defaults to FALSE, but If set to TRUE, 
the interactive prompt will be skipped and the updates will proceed automatically.}

\item{...}{Other arguments passed to either \code{\link[BiocManager:install]{BiocManager::install}} or \code{\link[remotes:install_github]{remotes::install_github}}}
}
\description{
This will check to see if all cytoverse packages (and optionally, their
dependencies) are up-to-date and installl those that are behind. The \code{repo} argument
will determine whether to update from Bioconductor or GitHub and the \code{pkgs} argument
allows for specification of a subset of packages to update.
}
\details{
This utilizes either \code{\link[BiocManager:install]{BiocManager::install}} for Bioconductor or \code{\link[remotes:install_github]{remotes::install_github}}
to check for package updates to install. As such, additional arguments can be passed to either of those methods.
Please see the examples for the most typical use cases. \code{If repo="bioconductor"} is used and \code{BiocManager} is unavailable, it will be installed.
}
\examples{
\dontrun{
# Default behavior. Update all cytoverse packages from
# appropriate Bioconductor release. Asks for approval of proposed updates.
cytoverse_update()

# Same as above, but do not ask for approval before proceeding with installation
cytoverse_update(upgrades_approved = TRUE)

# Still using appropriate Bioconductor release, but limiting to just checking
# a few packages for updates. check_versions = FALSE makes the method attempt to
# update all packages, relying on commit SHA to determine if packages are up-to-date
cytoverse_update(pkgs = c("flowWorkspace", "CytoML"), check_versions = FALSE)

# Specify that you would like to use the current development branches from Bioconductor
cytoverse_update(version = "devel")

# Override default repository (Bioconductor) to instead install from GitHub
cytoverse_update(repo = "github")

# An example showing extra arguments being passed down to install_github.
# Only update cytoqc and all of its dependencies (including from "Suggests" and "Enhances" sections)
cytoverse_update(repo = "github", pkgs = "cytoqc", dependencies = TRUE)

}
}
