% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{cytoverse_update}
\alias{cytoverse_update}
\title{Update cytoverse packages}
\usage{
cytoverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
cytoverse packages.}
}
\description{
This will check to see if all cytoverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
cytoverse_update()
}
}
